Refactor error hints into SRP modules
