+++ crates/tokmd/src/error_hints/edit_distance.rs
@@ -1,0 +1,32 @@
+pub(super) fn levenshtein(a: &str, b: &str) -> usize {
+    let a_chars: Vec<char> = a.chars().collect();
+    let b_chars: Vec<char> = b.chars().collect();
+
+    if a_chars.is_empty() {
+        return b_chars.len();
+    }
+    if b_chars.is_empty() {
+        return a_chars.len();
+    }
+
+    let mut d = vec![vec![0; b_chars.len() + 1]; a_chars.len() + 1];
+
+    for (i, row) in d.iter_mut().enumerate().take(a_chars.len() + 1) {
+        row[0] = i;
+    }
+    for (j, item) in d[0].iter_mut().enumerate().take(b_chars.len() + 1) {
+        *item = j;
+    }
+
+    for i in 1..=a_chars.len() {
+        for j in 1..=b_chars.len() {
+            let cost = usize::from(a_chars[i - 1] != b_chars[j - 1]);
+            d[i][j] = std::cmp::min(
+                std::cmp::min(d[i - 1][j] + 1, d[i][j - 1] + 1),
+                d[i - 1][j - 1] + cost,
+            );
+        }
+    }
+
+    d[a_chars.len()][b_chars.len()]
+}
